KeepFit Static Version
Shneiderman's "Eight Golden Rules of Interface Design", HTML 5, CSS 3, Bootstrap 5, and JavaScript were used to design this static prototype. The prototype is the teachers’ and instructors’ interface (UI) of the web version of a mobile fitness app which allows teachers and instructors track the activity of students using wrist-worn devices.